requirements.txt
fastapi
uvicorn[standard]
litellm
python-dotenv
pyyaml
orjson
pydantic
pydantic-settings
httpx
